+++ src/api/video_decoder.rs
@@ -493,14 +493,6 @@ impl VideoDecoderFactory {
 
     pub fn create(
         &self,
-        env: &crate::Environment,
-        format: &SdpVideoFormat,
-    ) -> Option<VideoDecoder> {
-        self.create_from_ref(env.as_ref(), format.as_ref())
-    }
-
-    pub fn create_from_ref(
-        &self,
         env: EnvironmentRef<'_>,
         format: SdpVideoFormatRef<'_>,
     ) -> Option<VideoDecoder> {